WIP Improvements - Fixes a bit more intrinsics - Extract helpers and operator modules - Fix disjoint warning for !== - Remove intermediate operation
