+++ checker/src/features/operations/mod.rs
@@ -1,0 +1,12 @@
+mod equality;
+mod logical;
+mod mathematical_bitwise;
+mod unary;
+
+pub use equality::{
+	evaluate_equality_inequality_operation, is_null_or_undefined, CanonicalEqualityAndInequality,
+	EqualityAndInequality, EqualityAndInequalityResultKind,
+};
+pub use logical::{evaluate_logical_operation_with_expression, LogicalOperator};
+pub use mathematical_bitwise::{evaluate_mathematical_operation, MathematicalOrBitwiseOperation};
+pub use unary::{evaluate_unary_operator, UnaryOperation};
