+++ checker/src/features/template_literal.rs
@@ -144,7 +144,7 @@ where
 				checking_data.types.new_constant_type(Constant::String(static_part.to_owned()));
 			let result = super::operations::evaluate_mathematical_operation(
 				acc,
-				crate::features::operations::MathematicalAndBitwise::Add,
+				crate::features::operations::MathematicalOrBitwiseOperation::Add,
 				lhs,
 				environment,
 				&mut checking_data.types,
@@ -164,7 +164,7 @@ where
 			);
 			let result = super::operations::evaluate_mathematical_operation(
 				acc,
-				crate::features::operations::MathematicalAndBitwise::Add,
+				crate::features::operations::MathematicalOrBitwiseOperation::Add,
 				rhs,
 				environment,
 				&mut checking_data.types,
@@ -184,7 +184,7 @@ where
 				checking_data.types.new_constant_type(Constant::String(final_part.to_owned()));
 			let result = super::operations::evaluate_mathematical_operation(
 				acc,
-				crate::features::operations::MathematicalAndBitwise::Add,
+				crate::features::operations::MathematicalOrBitwiseOperation::Add,
 				value,
 				environment,
 				&mut checking_data.types,
