+++ checker/src/options.rs
@@ -53,6 +53,11 @@ pub struct TypeCheckOptions {
 
 	pub measure_time: bool,
 
+	/// Enables two things:
+	/// - range and modulo class inference
+	/// - modifications to ranges and classes based on operations
+	pub advanced_number_intrinsics: bool,
+
 	/// Printing internal diagnostics in dts
 	pub debug_dts: bool,
 }
@@ -78,6 +83,7 @@ impl Default for TypeCheckOptions {
 			measure_time: false,
 			debug_dts: false,
 			extra_syntax: true,
+			advanced_number_intrinsics: false,
 		}
 	}
 }
