+++ checker/src/synthesis/type_annotations.rs
@@ -11,13 +11,11 @@ use crate::{
 	features::objects::ObjectBuilder,
 	types::{
 		generics::generic_type_arguments::GenericArguments,
+		generics::ExplicitTypeArguments,
+		helpers::{ArrayItem, Counter},
+		intrinsics::{self, distribute_tsc_string_intrinsic},
 		properties::{PropertyKey, PropertyValue, Publicity},
 		Constant, Constructor, PartiallyAppliedGenerics, PolyNature, Type, TypeId,
-	},
-	types::{
-		generics::ExplicitTypeArguments,
-		intrinsics::{self, distribute_tsc_string_intrinsic},
-		ArrayItem, Counter,
 	},
 	CheckingData, Map,
 };
@@ -490,7 +488,7 @@ pub fn synthesise_type_annotation<T: crate::ReadFromFS>(
 						items.push((pos, ArrayItem::Optional(annotation_ty)));
 					}
 					TupleElementKind::Spread => {
-						let slice = crate::types::as_slice(
+						let slice = crate::types::helpers::as_slice(
 							annotation_ty,
 							&checking_data.types,
 							environment,
@@ -521,15 +519,15 @@ pub fn synthesise_type_annotation<T: crate::ReadFromFS>(
 			);
 
 			for (ty_position, item) in items {
+				use crate::types::helpers::ArrayItem;
+
 				let value = match item {
-					crate::types::ArrayItem::Member(item_ty) => PropertyValue::Value(item_ty),
-					crate::types::ArrayItem::Optional(item_ty) => {
-						PropertyValue::ConditionallyExists {
-							condition: TypeId::OPEN_BOOLEAN_TYPE,
-							truthy: Box::new(PropertyValue::Value(item_ty)),
-						}
-					}
-					crate::types::ArrayItem::Wildcard(on) => {
+					ArrayItem::Member(item_ty) => PropertyValue::Value(item_ty),
+					ArrayItem::Optional(item_ty) => PropertyValue::ConditionallyExists {
+						condition: TypeId::OPEN_BOOLEAN_TYPE,
+						truthy: Box::new(PropertyValue::Value(item_ty)),
+					},
+					ArrayItem::Wildcard(on) => {
 						crate::utilities::notify!("found wildcard");
 						let after = idx.into_type(&mut checking_data.types);
 
@@ -723,7 +721,8 @@ pub fn synthesise_type_annotation<T: crate::ReadFromFS>(
 					checking_data.types.register_type(Type::Constructor(
 						crate::types::Constructor::BinaryOperator {
 							lhs: acc,
-							operator: crate::features::operations::MathematicalAndBitwise::Add,
+							operator:
+								crate::features::operations::MathematicalOrBitwiseOperation::Add,
 							rhs: lhs,
 							result: TypeId::STRING_TYPE,
 						},
@@ -740,7 +739,7 @@ pub fn synthesise_type_annotation<T: crate::ReadFromFS>(
 				};
 				let constructor = crate::types::Constructor::BinaryOperator {
 					lhs: acc,
-					operator: crate::features::operations::MathematicalAndBitwise::Add,
+					operator: crate::features::operations::MathematicalOrBitwiseOperation::Add,
 					rhs,
 					result: TypeId::STRING_TYPE,
 				};
@@ -756,7 +755,8 @@ pub fn synthesise_type_annotation<T: crate::ReadFromFS>(
 					checking_data.types.register_type(Type::Constructor(
 						crate::types::Constructor::BinaryOperator {
 							lhs: acc,
-							operator: crate::features::operations::MathematicalAndBitwise::Add,
+							operator:
+								crate::features::operations::MathematicalOrBitwiseOperation::Add,
 							rhs: lhs,
 							result: TypeId::STRING_TYPE,
 						},
