+++ checker/src/types/properties/assignment.rs
@@ -9,8 +9,9 @@ use crate::{
 	types::{
 		calling::{CallingDiagnostics, CallingOutput, SynthesisedArgument},
 		get_constraint,
+		helpers::tuple_like,
 		logical::{BasedOnKey, Logical, LogicalOrValid, NeedsCalculation},
-		tuple_like, Constructor, GenericChain, PartiallyAppliedGenerics, TypeStore,
+		Constructor, GenericChain, PartiallyAppliedGenerics, TypeStore,
 	},
 	Environment, Type, TypeId,
 };
